adapter: extract validate_read_then_write_dependencies for reuse
We will later use this for the frontend read-then-write implementation.
